applications/portal/backend/requirements.txt
uritemplate==4.1.1
psycopg2-binary
django-dry-rest-permissions==1.2.0
Pillow==9.0.1
bg-atlasapi==1.0.2

djangorestframework==3.13.1
